requirements.txt
beautifulsoup4==4.12.3
Django==5.0.3
html5lib==1.1
httpx==0.27.0
sh40-slate==0.7.0
sh40-zenith==0.6.0
watchfiles==0.21.0
